# business/docs/swagger.yaml
basePath: /api/v1
...
host: localhost:8082
info:
  contact:
    email: dev@watink.com.br
    name: Watink Dev
  description: API do backend Go — atendimento e automação WhatsApp.
  license:
    name: Proprietary
    url: https://watink.com.br
  termsOfService: https://watink.com.br/terms
  title: Watink Business API
  version: 2.0.0
paths:
...
  /knowledge-bases:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              additionalProperties: true
              type: object
            type: array
      security:
      - BearerAuth: []
      summary: Listar bases de conhecimento
      tags:
      - knowledge-base
    post:
      consumes:
      - application/json
      parameters:
      - description: Dados da base
        in: body
        name: body
        required: true
        schema:
          additionalProperties: true
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Criar base de conhecimento
      tags:
      - knowledge-base
  /knowledge-bases/{knowledgeBaseId}:
    delete:
      parameters:
      - description: ID da base
        in: path
        name: knowledgeBaseId
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Remover base de conhecimento
      tags:
      - knowledge-base
    get:
      parameters:
      - description: ID da base
        in: path
        name: knowledgeBaseId
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Detalhar base de conhecimento
      tags:
      - knowledge-base
    put:
      consumes:
      - application/json
      parameters:
      - description: ID da base
        in: path
        name: knowledgeBaseId
        required: true
        type: integer
      - description: Campos a atualizar
        in: body
        name: body
        required: true
        schema:
          additionalProperties: true
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Atualizar base de conhecimento
      tags:
      - knowledge-base
  /knowledge-bases/{knowledgeBaseId}/sources:
    post:
      consumes:
      - application/json
      parameters:
      - description: ID da base
        in: path
        name: knowledgeBaseId
        required: true
        type: integer
      - description: Dados da fonte
        in: body
        name: body
        required: true
        schema:
          additionalProperties: true
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Adicionar fonte na base
      tags:
      - knowledge-base
  /knowledge-bases/{knowledgeBaseId}/sources/{sourceId}:
    delete:
      parameters:
      - description: ID da base
        in: path
        name: knowledgeBaseId
        required: true
        type: integer
      - description: ID da fonte
        in: path
        name: sourceId
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Remover fonte da base
      tags:
      - knowledge-base
  /messages/{ticketId}:
    get:
      parameters:
      - description: ID do ticket
        in: path
        name: ticketId
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              additionalProperties: true
              type: object
            type: array
      security:
      - BearerAuth: []
      summary: Listar mensagens do ticket
      tags:
      - messages
    post:
      consumes:
      - application/json
      description: Envia mensagem via engine WhatsApp através do RabbitMQ
      parameters:
      - description: ID do ticket
        in: path
        name: ticketId
        required: true
        type: integer
      - description: Conteúdo da mensagem
        in: body
        name: body
        required: true
        schema:
          additionalProperties: true
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Enviar mensagem
      tags:
      - messages
...
securityDefinitions:
  BearerAuth:
    description: 'Bearer JWT token. Formato: "Bearer <token>"'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"